fix: Pass gitEnv to checkout commands
Passes the git environment variables (containing the PAT) to the `git checkout` commands in the `ensureBranch` method.

This is necessary to authenticate when checking out branches in a "blobless" clone, as Git needs to download the file contents (blobs) from the promisor remote. This fixes the "Cannot switch to target branch" error.